@@ -22,19 +22,27 @@
     "start": "npm run build -- -w",
     "test": "node tests/rules/*.js",
     "pret": "prettier --write .",
-    "lint": "eslint src/",
-    "eg": "eslint --rulesdir lib/rules examples/"
+    "lint": "eslint .",
+    "eg": "cd examples && eslint ."
   },
   "peerDependencies": {
     "eslint": ">=3.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.19.3",
-    "@babel/core": "^7.19.3",
-    "@babel/preset-env": "^7.19.4",
-    "eslint": "^8.25.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "prettier": "^2.7.1"
+    "@babel/cli": "^7.25.7",
+    "@babel/core": "^7.25.7",
+    "@babel/preset-env": "^7.25.7",
+    "@eslint/compat": "^1.1.1",
+    "@eslint/js": "^9.11.1",
+    "eslint": "^9.11.1",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "globals": "^15.10.0",
+    "prettier": "^3.3.3"
+  },
+  "overrides": {
+    "eslint-plugin-react-hooks": {
+      "eslint": "$eslint"
+    }
   }
 }
